+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/header/cover_editor.dart
